refactor(CLI): got parser help text from type hints from Config Struct field Meta descriptions
